Store the verbatim score as a number the reader can actually recognize
Bugbot: the column was `numeric`, and PostgREST serializes numeric as a STRING
to preserve arbitrary precision. verbatimBadgeState checked typeof === "number",
so every stored score would have failed that check and the badge would have
silently never rendered. Nothing would have failed; it simply would never have
worked, which is the hardest kind of bug to notice in a badge.

Fixed at the source: the column is double precision, which PostgREST returns as
a real JSON number. A 0-1 ratio never needed arbitrary precision, and numeric is
the only reason a string was ever involved.

The reader also accepts a string now. That is belt and braces rather than
redundancy: the failure mode of rejecting one is invisible, so the cost of
tolerating it is nil next to the cost of being wrong again.

Writing that test caught a second edge my own fix had introduced. Number("") is
0, not NaN, so a blank string rendered "script 0%" and claimed the assistant
read NONE of its script. An empty value means "no score", and showing nothing
is the honest answer.

Also in this commit, from the earlier build failure: the two badge decisions
moved out of src/lib/db/voice-transcripts.ts, which imports the server-only
Supabase client, into src/lib/voice/transcript-badges.ts. tsc and eslint both
pass on a server import from a client component; only `next build` walks the
bundle boundary, and that is the check this needed.

+++ supabase/migrations/20260822100237_voice_transcript_amd_columns.sql
@@ -19,6 +19,12 @@
 --   voicemail_verbatim_score  how closely the spoken message matched the
 --                             approved script (see _shared/voice_verbatim.ts).
 --                             NULL until a voicemail is actually left.
+--                             double precision, NOT numeric: PostgREST
+--                             serializes numeric as a STRING to preserve
+--                             arbitrary precision, so a reader doing a
+--                             typeof === "number" check would silently never
+--                             see a score. A 0-1 ratio has no need of that
+--                             precision.
 --
 -- grants: none (voice_call_transcripts): existing table with its own policies;
 -- adding columns does not change its Data API exposure.
@@ -26,7 +32,7 @@
 alter table public.voice_call_transcripts
   add column if not exists answering_machine_result text,
   add column if not exists voicemail_left boolean not null default false,
-  add column if not exists voicemail_verbatim_score numeric;
+  add column if not exists voicemail_verbatim_score double precision;
 
 comment on column public.voice_call_transcripts.answering_machine_result is
   'Answering-machine detection verdict for this leg (human / machine / unknown), or NULL when AMD was not requested. Premium AMD reports human_residence / human_business rather than a bare "human", so the raw value is normalized before it is stored; see _shared/voice_amd.ts.';
